% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{coef.fs}
\alias{coef.fs}
\title{Coefficient function for fs object.}
\usage{
\method{coef}{fs}(object, s, ...)
}
\arguments{
\item{object}{The fs path object, as produced by the fs function.}

\item{s}{The step (or vector of steps) of the path at which coefficients
should be computed.  Can be fractional, in which case interpolation is
performed.}

\item{...}{Other arguments (currently not used).}
}
\description{
Compute coefficients at a particular step of the forward stepwise path.
}

