% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{fs}
\alias{fs}
\title{Forward stepwise regression.}
\usage{
fs(x, y, maxsteps = 2000, intercept = TRUE, normalize = TRUE,
  verbose = FALSE)
}
\arguments{
\item{x}{Matrix of predictors, of dimension (say) n x p.}

\item{y}{Vector of responses, of length (say) n.}

\item{maxsteps}{Maximum number of steps of the forward stepwise path to
compute. Default is 2000.}

\item{intercept, normalize}{Should an intercept be included in the regression
model? Should the predictors be normalized before computing the path?
Default is TRUE for both.}

\item{verbose}{Should intermediate progress be printed out? Default is FALSE.}
}
\value{
A list with the following components:
  \itemize{
  \item action, sign: vectors that give the index of the variable added at 
  each step, and the sign of this variable's correlation with the residual
  upon entry df
  \item df: vector that gives the (naive) degrees of freedom of the model 
  at each step, i.e., the number of active predictor variables (+ 1 if there
  is an intercept included)
  \item completepath: a boolean indicating whether the forward stepwise path
  was run to completion (as opposed to being stopped early because the max
  number of steps was achieved)
  \item bls: if the complete path was computed, this is a vector that gives
  the least squares coefficients of the full regression model
  \item x, y: the passed x and y
  \item bx, by: the means of the columns of x, and the mean of y
  \item intercept, normalize: the passed values for intercept and normalize
  \item call: the call to fs()
  }
}
\description{
Compute the forward stepwise regression path.
}
\details{
This function implements forward stepwise regression, adding the
  predictor at each step that maximizes the absolute correlation between the
  predictors---once orthogonalized with respect to the current model---and
  the residual. This entry criterion is standard, and is equivalent to
  choosing the variable that achieves the biggest drop in RSS at each step;
  it is used, e.g., by the step function in R. Note that, for example,
  the lars package implements a stepwise option (with type="step"),
  but uses a (mildly) different entry criterion, based on maximal absolute
  correlation between the original (non-orthogonalized) predictors and the 
  residual.
}
\author{
Ryan Tibshirani, Trevor Hastie, Rob Tibshirani
}

