% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Lasso and friends.}
\usage{
lasso(x, y, alpha = 1, nrelax = 10, nlambda = 50,
  lambda.min.ratio = ifelse(nrow(x) < ncol(x), 0.01, 1e-04), lambda = NULL,
  intercept = TRUE, standardize = TRUE)
}
\description{
This is just a simple wrapper function around the
  \code{\link{glmnet}} function in the R package of the same name. Its 
  purpose is to provide a version where the associated coef and predict
  methods always produce coefficients and predictions at exactly nlambda
  values, by default. (The \code{\link{glmnet}} function may produce a
  path with less than nlambda lambda values, depending on the data.)
}
\details{
Compute the lasso, ridge regression, or elastic net solutions in regression.
}

