% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.from.file}
\alias{plot.from.file}
\title{Plot the results over several simulation settings.}
\usage{
\method{plot}{from.file}(file.list, row = c("beta", "rho", "snr"),
  col = c("rho", "beta", "snr"), method.nums = NULL, method.names = NULL,
  what = c("error", "risk", "prop", "F", "nonzero"), rel.to = NULL,
  tuning = c("validation", "oracle"), type = c("ave", "med"), std = TRUE,
  lwd = 1, pch = 19, main = NULL, ylim = NULL,
  legend.pos = c("bottom", "right", "top", "left", "none"),
  make.pdf = FALSE, fig.dir = ".", file.name = "sim", w = 8, h = 10)
}
\arguments{
\item{file.list}{Vector of strings that point to saved sim objects (each
object produced by a call to \code{\link{sim.master}}).}

\item{row, col}{One of "beta", "rho", or "snr", indicating the variables for
the rows and columns of plotting grid; note that row and col must be
different.  Default is row="beta" and col="rho", so that the plotting grid
displays the metric specified in what (default is relative test error, see
below) versus the SNR, in a plotting grid with the coefficient types across
the rows, and the correlation levels across the columns.}

\item{method.nums}{The indices of the methods that should be plotted. Default
is NULL, in which case all methods are plotted.}

\item{method.names}{The names of the methods that should be plotted. Default
is NULL, in which case the names are extracted from the sim objects.}

\item{what}{One of "error", "risk", "prop", or "nonzero", indicating whether
to plot the relative test error, test proportion of variance explained, or
number of nonzeros, for each method across the given SNR levels. When what
is "prop", the x-axis is population proportion of variance explained,
instead of SNR. Default is "error".}

\item{rel.to}{An index for a base method: when the what argument is "error"
or "risk", the rel.to argument specifies the method with respect to which
the relative error or relative risk is calculated. The default value of
NULL means different things when what is equal to "error" and "risk". In
the former case, the test error is calculated relative to the oracle; in
the latter, the risk is calculated relative to the null model.}

\item{tuning}{Either "validation" or "oracle", indicating whether the tuning
parameter for each method should be chosen according to minimizing
validation error, or according to minimizing test error. Default is
"validation".}

\item{type}{Either "ave" or "med", indicating whether the average or median
of the test error (or number of nonzeros, if what is "nonzero") should be
displayed. Default is "ave".}

\item{std}{Should standard errors be displayed (in parantheses)? When type
is set to "med", the median absolute deviations are shown in place of the
standard errors. Default is TRUE.}

\item{lwd, pch, main, ylim, legend.pos}{graphical parameters.}

\item{make.pdf}{Should a pdf be produced? Default is FALSE.}

\item{fig.dir, file.name}{The figure directory and file name to use, only
when make.pdf is TRUE. Defaults are "." and "sim". (An extension of "pdf"
is always appended to the given file name.)}

\item{w, h}{the width and height (in inches) for the plot, used only when
make.pdf is TRUE. Defaults are 8 and 10, respectively.}
}
\description{
Plot results over several sets of simulations saved in files, where the same
  methods are run over different simulations settings.
}

