% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{plot.sim}
\alias{plot.sim}
\title{Plot function for sim object.}
\usage{
\method{plot}{sim}(x, method.nums = 1:length(x$err.train.ave),
  method.names = NULL, type = c("ave", "med"), std = TRUE, cols = 1:8,
  main = NULL, cex.main = 1.25, legend.pos = c("topright"),
  make.pdf = FALSE, fig.dir = ".", file.name = "sim", w = 6, h = 6,
  mar = NULL)
}
\arguments{
\item{x}{The sim object.}

\item{method.nums}{the indices of the methods that should be plotted. Default
is to 1:length(x$err.rel.ave), which plots all methods.}

\item{method.names}{the names of the methods that should be plotted. Default
is NULL, in which case the names are extracted from the sim object.}

\item{type}{Either "ave" or "med", indicating whether the average or median
of the relative test error metric should be displayed. Default is "ave".}

\item{std}{Should standard errors be displayed (in parantheses)? When type
is set to "med", the median absolute deviations are shown in place of the
standard errors. Default is TRUE.}

\item{cols, main, cex.main, legend.pos}{graphical parameters.}

\item{make.pdf}{Should a pdf be produced? Default is FALSE.}

\item{fig.dir, file.name}{The figure directory and file name to use, only
when make.pdf is TRUE. Defaults are "." and "sim". (An extension of "pdf"
is always appended to the given file name.)}

\item{w, h}{the width and height (in inches) for the plot, used only when
make.pdf is TRUE. Defaults are 6 for both.}

\item{mar}{the margins to use for the plot. Default is NULL, in which case
the margins are set automatically (depending on whether not main is NULL).}
}
\description{
Plot the results of a set of simulations, stored in an object of class sim
  (produced by \code{\link{sim.master}}).
}

