% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{print.sim}
\alias{print.sim}
\title{Print function for sim object.}
\usage{
\method{print}{sim}(x, what = c("error", "risk"), type = c("ave", "med"),
  std = TRUE, digits = 3, ...)
}
\arguments{
\item{x}{The sim object.}

\item{what}{Either "error" or "risk", indicating whether the relative test
error or relative risk should be displayed. Default is "error".}

\item{type}{Either "ave" or "med", indicating whether the average or median
of the relative test error metric should be displayed. Default is "ave".}

\item{std}{Should standard errors be displayed (in parantheses)? When type
is set to "med", the median absolute deviations are shown in place of the
standard errors. Default is TRUE.}

\item{digits}{Number of digits to display. Default is 3.}

\item{...}{Other arguments (currently not used).}
}
\description{
Summarize and print the results of a set of simulations, stored in an object
  of class sim (produced by \code{\link{sim.master}}).
}

