% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.xy}
\alias{sim.xy}
\title{Predictors and responses generation.}
\usage{
sim.xy(n, p, nval, ntest, rho = 0, s = 5, beta.type = 1, snr = 1)
}
\arguments{
\item{n, p}{The number of training observations, and the number of predictors.}

\item{nval, ntest}{The number of validation observations, and the number of
testing observations.}

\item{rho}{Parameter that drives pairwise correlations of the predictor
variables; specifically, predictors i and j will have pairwise correlation
rho^abs(i-j). Default is 0.}

\item{s}{number of nonzero coefficients in the underlying regression model.
Default is 5. (Ignored if beta.type is 4, in which case the number of
nonzero coefficients is 6; and if beta.type is 5, it is interpreted as a
the number of strongly nonzero coefficients in a weak sparsity model.)}

\item{beta.type}{Integer taking values in between 1 and 5, used to specify
the pattern of nonzero coefficients in the underlying regression model; see
details below. Default is 1.}

\item{snr}{Desired signal-to-noise ratio; the variance of the errors will be
set accordingly. Default is 1.}
}
\value{
A list with the following components: x, y, xval, yval, xtest, ytest,
  mutest, beta, and sigma.
}
\description{
Generate a predictor matrix x, and response vector y, following a specified
  setup.  Actually, three pairs of predictors and responses are generated:
  one for training, one for validation, and one for testing.
}
\details{
The predictors are normal with covariance sigma^2 * Sigma, where
  sigma^2 is set according to the desired signal-to-noise ratio, and Sigma
  has (i,j)th entry rho^abs(i-j). The first 4 options for the nonzero pattern
  of the underlying regression coefficients beta follow the simulation setup
  in Bertsimas, King, and Mazumder (2016), and the last is a weak sparsity
  option:
  \itemize{
  \item 1: beta has s components of 1, occurring at (roughly) equally-spaced
     indices in between 1 and p
  \item 2: beta has its first s components equal to 1
  \item 3: beta has its first s components taking nonzero values, where the
      decay in a linear fashion from 10 to 0.5
  \item 4: beta has its first 6 components taking the nonzero values -10,-6,
      -2,2,6,10
  \item 5: beta has its first s components equal to 1, and the rest decaying
      to zero at an exponential rate
  }
}
\author{
Trevor Hastie, Rob Tibshirani, Ryan Tibshirani
}
\references{
Simulation setup based on "Best subset selection via a modern
  optimization lens" by Dimitris Bertsimas, Angela King, and Rahul Mazumder,
  Annals of Statistics, 44(2), 813-852, 2016.
}

