% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{coef.fs}
\alias{coef.fs}
\title{Coefficient function for fs object.}
\usage{
\method{coef}{fs}(object, s, ...)
}
\arguments{
\item{object}{The fs object, as produced by the fs function.}

\item{s}{The step (or vector of steps) of the path at which coefficients
should be computed. Can be fractional, in which case interpolation is
performed. If missing, then the default is use all steps of the passed
fs object.}

\item{...}{Other arguments (currently not used).}
}
\description{
Compute coefficients at a particular step of the forward stepwise path.
}
\details{
Note that at s = 1, there is one nonzero coefficient, at
  s = 2, there are two nonzero coefficients, etc. (This differs from the
  parametrization used in the \code{coef.fs} function in the R package
  \code{selectiveInference}, as the latter function delivers s-1 nonzero
  coefficients at step s, and was written to be consistent with the
  natural parametrization for the least angle regression path.)
}

