% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Lasso and friends.}
\usage{
lasso(x, y, alpha = 1, nrelax = 1, nlambda = 50,
  lambda.min.ratio = ifelse(nrow(x) < ncol(x), 0.01, 1e-04), lambda = NULL,
  intercept = TRUE, standardize = TRUE)
}
\arguments{
\item{nrelax}{The number of interpolations to produce between the lasso
(or ridge, or elastic net) solution, at each value of lambda, and the
least squares coefficients on the corresponding active set. (The number
of interpolations counts the endpoints inclusively, i.e., the lasso and
least squares solutions.) Default is 1, which means that only the lasso
solution is considered (no strict relaxations), at each value of lambda.}
}
\description{
This is just a simple wrapper function around the
  \code{\link{glmnet}} function in the R package of the same name. Its
  purpose is twofold: (i) to provide functionality where the associated coef
  and predict methods always produce coefficients and predictions at exactly
  nlambda values, by default (the \code{\link{glmnet}} function may produce a
  path with less than nlambda lambda values, depending on the data); and (ii)
  to provide relaxed versions of the lasso (or ridge regression, or elastic
  net) solutions, defined by interpolating in between each solution and the
  least squares coefficients on the corresponding active set. The number of
  interpolations is governed by the nrelax argument; all other arguments are
  the same as in \code{\link{glmnet}}.
}
\details{
Compute the lasso, ridge regression, or elastic net solutions in regression.
}
\author{
Trevor Hastie, Robert Tibshirani, Ryan Tibshirani
}

