% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.many.sims}
\alias{plot.many.sims}
\title{Plot the results over several simulation settings.}
\usage{
plot.many.sims(file.list, grouping, snr.vec, method.nums = NULL,
  method.names = NULL, what = c("error", "prop", "nonzero"),
  tuning = c("validation", "oracle"), type = c("ave", "med"), std = TRUE,
  fig.dir = ".", file.name = NULL, w = 6, h = 6, mar = NULL,
  cols = 1:8, main = NULL, cex.main = 1.25, log = ifelse(what == "prop",
  "", "x"), legend.pos = "bottomright", tex.dir = NULL)
}
\arguments{
\item{file.list}{vector of strings that point to saved sim objects (each
object produced by a call to \code{\link{sim.master}}).}

\item{grouping}{integer or factor vector indicating the grouping to use for
the simulations. Within each group, the relative test error achieved by
each method is plotted across the available SNR levels.}

\item{snr.vec}{Vector giving the SNR levels considered within each group.}

\item{method.nums}{the indices of the methods that should be plotted. Default
is NULL, in which case all methods are plotted.}

\item{method.names}{the names of the methods that should be plotted. Default
is NULL, in which case the names are extracted from the sim objects.}

\item{what}{one of "error", "prop", or "nonzero", indicating whether to plot
the relative test error, test proportion of variance explained, or number
of nonzeros, for each method across the given SNR levels. When what is
"prop", the x-axis is population proportion of variance explained, instead
of SNR. Default is "error".}

\item{tuning}{one of "validation" or "oracle", indicating whether the tuning
parameter for each method should be chosen according to minimizing
validation error, or according to minimizing test error. Default is
"validation".}

\item{type}{Either "ave" or "med", indicating whether the average or median
of the test error (or number of nonzeros, if what is "nonzero") should be
displayed. Default is "ave".}

\item{std}{Should standard errors be displayed (in parantheses)? When type
is set to "med", the median absolute deviations are shown in place of the
standard errors. Default is TRUE.}

\item{fig.dir}{The figure directory to use. Default is ".".}

\item{file.name}{Vector of strings, giving the file names to use for the 
saved figures. Default is NULL, in which case the names "sim1", "sim2",
etc. are used. (Extensions of "pdf" are always appended to the given file
names.)}

\item{w, h}{the width and height (in inches) for the plots. Defaults are 6 for
both.}

\item{mar}{the margins to use for the plots. Default is NULL, in which case
the margins are set automatically (depending on whether not main is NULL).}

\item{cols, main, cex.main, log, legend.pos}{graphical parameters.}

\item{tex.dir}{The latex directory to use. Default is NULL, which means that
no lines of tex (includegraphics statements, for the figures created) will
be produced.}
}
\description{
Plot results over several sets of simulations, where the same methods are run
  over different simulations settings.
}

