% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.xy}
\alias{sim.xy}
\title{Predictors and responses generation.}
\usage{
sim.xy(n, p, nval, rho = 0, s = 5, beta.type = 1, snr = 1)
}
\arguments{
\item{n, p}{The number of training observations, and the number of predictors.}

\item{nval}{The number of validation observations.}

\item{rho}{Parameter that drives pairwise correlations of the predictor
variables; specifically, predictors i and j have population correlation
rho^abs(i-j). Default is 0.}

\item{s}{number of nonzero coefficients in the underlying regression model.
Default is 5. (Ignored if beta.type is 4, in which case the number of
nonzero coefficients is 6; and if beta.type is 5, it is interpreted as a
the number of strongly nonzero coefficients in a weak sparsity model.)}

\item{beta.type}{Integer taking values in between 1 and 5, used to specify
the pattern of nonzero coefficients in the underlying regression model; see
details below. Default is 1.}

\item{snr}{Desired signal-to-noise ratio (SNR), i.e., var(mu)/sigma^2 where
mu is mean and sigma^2 is the error variance. The error variance is set so
that the given SNR is achieved. Default is 1.}
}
\value{
A list with the following components: x, y, xval, yval, Sigma, beta,
  and sigma.
}
\description{
Generate a predictor matrix x, and response vector y, following a specified
  setup.  Actually, two pairs of predictors and responses are generated:
  one for training, and one for validation.
}
\details{
The data model is: \eqn{Y \sim N(X\beta, \sigma^2 I)}.
  The predictor variables have covariance matrix Sigma, with (i,j)th entry
  rho^abs(i-j). The error variance sigma^2 is set according to the desired
  signal-to-noise ratio. The first 4 options for the nonzero pattern
  of the underlying regression coefficients beta follow the simulation setup
  in Bertsimas, King, and Mazumder (2016), and the 5th is a weak sparsity
  option:
  \itemize{
  \item 1: beta has s components of 1, occurring at (roughly) equally-spaced
     indices in between 1 and p
  \item 2: beta has its first s components equal to 1
  \item 3: beta has its first s components taking nonzero values, where the
      decay in a linear fashion from 10 to 0.5
  \item 4: beta has its first 6 components taking the nonzero values -10,-6,
      -2,2,6,10
  \item 5: beta has its first s components equal to 1, and the rest decaying
      to zero at an exponential rate
  }
}
\examples{
# Simulate some simple regression data with the first 5 coefficients
# being nonzero
set.seed(0)
n = 100
p = 20
xy.obj = sim.xy(n,p,nval=0,s=5,beta.type=2,snr=1)
x = xy.obj$x
y = xy.obj$y

# Run forward stepwise regression for 8 steps
fs.obj = fs(x,y,intercept=FALSE,maxsteps=8,verbose=TRUE)
fs.beta = coef(fs.obj)
fs.supp = apply(fs.beta != 0, 2, which)

# Solve best subset selection for 8 sparsity levels
bs.obj = bs(x,y,intercept=FALSE,k=0:8,verbose=TRUE)
bs.beta = coef(bs.obj)
bs.supp = apply(bs.beta != 0, 2, which)

# Compare supports of the solutions with 5 and 8 variables
fs.supp[[6]]; bs.supp[[6]]
fs.supp[[9]]; bs.supp[[9]]

# Predict on test data and record test error
ntest = 10000
xy.obj.test = sim.xy(ntest,p,nval=0,s=5,beta.type=2,snr=1)
xtest = xy.obj.test$x
ytest = xy.obj.test$y

fs.pred = predict(fs.obj,newx=xtest)
bs.pred = predict(bs.obj,newx=xtest)
colMeans((fs.pred - ytest)^2)
colMeans((bs.pred - ytest)^2)
}
\author{
Trevor Hastie, Rob Tibshirani, Ryan Tibshirani
}
\references{
Simulation setup based on "Best subset selection via a modern
  optimization lens" by Dimitris Bertsimas, Angela King, and Rahul Mazumder,
  Annals of Statistics, 44(2), 813-852, 2016.
}

